+++ apps/extension/src/projectInfo.ts
@@ -1,6 +1,6 @@
 export const PROJECT_INFO = {
   name: 'XShield',
-  version: '1.3.0',
+  version: '1.3.1',
   license: 'MIT',
   copyright: 'Copyright (c) 2026 XShield contributors',
   repositoryUrl: 'https://github.com/smthdagg/XShield',